+++ perl/FSM/Pipeline/HDLGenerator.pm
@@ -3838,10 +3838,10 @@ sub composition_signal_family_contexts ($self, $composition_plan, $signal_name, 
 
 sub build_composition_override_events ($self, $composition_plan, $structural_rtl_ir = undef) {
     my @events;
-    my @resolved_links = @{$composition_plan->resolved_links || []};
     my %same_name_endpoints;
     $structural_rtl_ir //= $self->build_composition_structural_rtl_ir($composition_plan);
     my $structural_rtl_ir_hash = ref($structural_rtl_ir) ? $structural_rtl_ir->as_hashref : {};
+    my @resolved_links = @{$structural_rtl_ir_hash->{resolved_links} || []};
 
     for my $instance (@{$structural_rtl_ir_hash->{instances} || []}) {
         for my $port (@{$instance->{interface_ports} || []}) {
@@ -3861,16 +3861,16 @@ sub build_composition_override_events ($self, $composition_plan, $structural_rtl
         next unless @same_name_candidates;
 
         my @touching_explicit_toplinks = grep {
-            (($_->origin_kind || '') eq 'declared_explicit_toplink')
+            ((($_->{origin_kind}) || '') eq 'declared_explicit_toplink')
             && (
-                (($_->source || '') eq $top_port->{name})
-                || (($_->target || '') eq $top_port->{name})
+                ((($_->{source}) || '') eq $top_port->{name})
+                || ((($_->{target}) || '') eq $top_port->{name})
             )
         } @resolved_links;
         next unless @touching_explicit_toplinks;
         my $example_link = $touching_explicit_toplinks[0];
-        my $source_context = $self->composition_provenance_endpoint_context($composition_plan, $example_link->source, $structural_rtl_ir);
-        my $target_context = $self->composition_provenance_endpoint_context($composition_plan, $example_link->target, $structural_rtl_ir);
+        my $source_context = $self->composition_provenance_endpoint_context($composition_plan, $example_link->{source}, $structural_rtl_ir);
+        my $target_context = $self->composition_provenance_endpoint_context($composition_plan, $example_link->{target}, $structural_rtl_ir);
 
         if (($top_port->{direction} || '') eq 'input') {
             my @compatible = grep {
@@ -3920,14 +3920,14 @@ sub build_composition_override_events ($self, $composition_plan, $structural_rtl
 
     my %seen_reexports;
     for my $resolved_link (@resolved_links) {
-        next unless ($resolved_link->origin_kind || '') eq 'inferred_internal_carrier_reexport_link';
-        my $top_port_name = $resolved_link->target;
+        next unless (($resolved_link->{origin_kind} || '') eq 'inferred_internal_carrier_reexport_link');
+        my $top_port_name = $resolved_link->{target};
         next if $seen_reexports{$top_port_name}++;
 
         push @events, {
             kind => 'explicit_top_output_reexports_internal_carrier',
             top_port_name => $top_port_name,
-            source => $resolved_link->source,
+            source => $resolved_link->{source},
             lane => $composition_plan->lane,
             top_port_context => $self->composition_provenance_endpoint_context($composition_plan, $top_port_name, $structural_rtl_ir),
             source_context => ($self->composition_signal_family_contexts($composition_plan, $top_port_name, 'output', $structural_rtl_ir)->[0]),
@@ -3943,7 +3943,7 @@ sub build_composition_block_events ($self, $composition_plan, $structural_rtl_ir
     my $structural_rtl_ir_hash = ref($structural_rtl_ir) ? $structural_rtl_ir->as_hashref : {};
     my %declared_top_ports = map { (($_->{name} || '') => $_) } @{$structural_rtl_ir_hash->{ports} || []};
     my @declared_links = @{$composition_plan->links || []};
-    my @resolved_links = @{$composition_plan->resolved_links || []};
+    my @resolved_links = @{$structural_rtl_ir_hash->{resolved_links} || []};
     my %port_groups;
 
     for my $instance (@{$structural_rtl_ir_hash->{instances} || []}) {
@@ -4014,13 +4014,13 @@ sub build_composition_block_events ($self, $composition_plan, $structural_rtl_ir
     my %internal_carrier_family;
     my %reexported_internal_carrier_family;
     for my $link (@resolved_links) {
-        my $raw_token = $link->raw_token || '';
+        my $raw_token = $link->{raw_token} || '';
         my ($family_name) = $raw_token =~ /^=implicit-internal:(\w+)$/;
         next unless defined $family_name;
 
-        if (($link->origin_kind || '') eq 'inferred_internal_carrier_link') {
+        if (($link->{origin_kind} || '') eq 'inferred_internal_carrier_link') {
             $internal_carrier_family{$family_name} = 1;
-        } elsif (($link->origin_kind || '') eq 'inferred_internal_carrier_reexport_link') {
+        } elsif (($link->{origin_kind} || '') eq 'inferred_internal_carrier_reexport_link') {
             $reexported_internal_carrier_family{$family_name} = 1;
         }
     }